# docs/api/admin-base-openapi.apifox.yaml
openapi: 3.0.3
info:
  title: admin_base API
  version: 2.0.0
  description: |
    Apifox 可导入的 admin_base 接口文档。接口以 `/admin-api` 为 context path。
    除 `/open/**` 外，其余接口需要 `Authorization: Bearer <token>`。
servers:
  - url: http://localhost:9999/admin-api
    description: local/dev server
security:
  - bearerAuth: []
...
paths:
...
  /role/all:
    post:
      tags: [Role]
      summary: 全部角色
      security: [{bearerAuth: []}]
      responses:
        "200":
          $ref: "#/components/responses/JsonSuccessOrError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /role/add:
    post:
      tags: [Role]
      summary: 新增角色
      security: [{bearerAuth: []}]
      requestBody:
        $ref: "#/components/requestBodies/AddRoleForm"
      responses:
        "200":
          $ref: "#/components/responses/JsonSuccessOrError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /role/delete:
    post:
      tags: [Role]
      summary: 删除角色
      security: [{bearerAuth: []}]
      requestBody:
        $ref: "#/components/requestBodies/RoleIdForm"
      responses:
        "200":
          $ref: "#/components/responses/JsonSuccessOrError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /role/update:
    post:
      tags: [Role]
      summary: 更新角色
      security: [{bearerAuth: []}]
      requestBody:
        $ref: "#/components/requestBodies/UpdateRoleForm"
      responses:
        "200":
          $ref: "#/components/responses/JsonSuccessOrError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /common/upload:
    post:
      tags: [Common]
      summary: 上传文件
      description: 需要权限 `sys:file:upload`。
      security: [{bearerAuth: []}]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          $ref: "#/components/responses/JsonSuccessOrError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /common/download:
    get:
      tags: [Common]
      summary: 通用下载
      description: 需要 `sys:file:download`；当 `delete=true` 时还需要 `sys:file:delete`。
      security: [{bearerAuth: []}]
      parameters:
        - {name: fileName, in: query, required: true, schema: {type: string}, example: sample_hello.txt}
        - {name: type, in: query, required: true, schema: {type: integer}, example: 0}
        - {name: delete, in: query, required: true, schema: {type: boolean}, example: false}
      responses:
        "200":
          description: 文件流；非法文件名时当前实现记录日志并返回空响应
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
  /common/download/resource2:
    get:
      tags: [Common]
      summary: 本地资源下载
      description: 需要权限 `sys:file:download`。
      security: [{bearerAuth: []}]
      parameters:
        - {name: resource, in: query, required: true, schema: {type: string}, example: sample_hello.txt}
      responses:
        "200":
          description: 文件流；非法文件名时当前实现记录日志并返回空响应
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    JsonSuccessOrError:
      description: 业务成功或业务错误。权限不足由全局异常处理返回 HTTP 200 + code=401。
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/JsonResponse"
              - $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            success:
              value: {code: 200, msg: 成功, data: null}
            forbidden:
              value: {code: 401, msg: 没有权限访问, data: null}
            validation:
              value: {code: 500, msg: page 不能为空, data: null}
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            code: 1003
            msg: 请先登录
            data: null
  requestBodies:
...
    AddRoleForm:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            required: [roleName, permissionIds]
            properties:
              roleName: {type: string}
              note: {type: string}
              permissionIds: {type: array, items: {type: integer}}
          example: {roleName: ROLE_TEST, note: test role, permissionIds: [1, 2]}
    RoleIdForm:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            required: [roleId]
            properties:
              roleId: {type: integer}
          example: {roleId: 2}
    UpdateRoleForm:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            required: [roleId, roleName, permissionIds]
            properties:
              roleId: {type: integer}
              roleName: {type: string}
              note: {type: string}
              permissionIds: {type: array, items: {type: integer}}
          example: {roleId: 2, roleName: ROLE_TEST, note: updated, permissionIds: [1, 2]}
...
  schemas:
    JsonResponse:
      type: object
      required: [code, msg]
      properties:
        code: {type: integer, example: 200}
        msg: {type: string, example: 成功}
        data: {}
    ErrorEnvelope:
      type: object
      required: [code, msg]
      properties:
        code: {type: integer, example: 500}
        msg: {type: string, example: 参数错误}
        data: {nullable: true}
...