# src/main/java/com/admin/base/infrastructure/controller/CommonController.java
package com.admin.base.infrastructure.controller;

import com.admin.base.shared.constant.ResponseCode;
import com.admin.base.shared.api.JsonResponse;
import com.admin.base.infrastructure.config.common.SysConfig;
import com.admin.base.shared.constant.DownloadType;
import com.admin.base.shared.exception.BusinessException;
import com.admin.base.shared.util.StringUtils;
import com.admin.base.shared.util.file.FileUploadUtils;
import com.admin.base.shared.util.file.FileUtils;
import jakarta.servlet.http.HttpServletResponse;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.MediaType;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

/**
 * @author ZXX
 * @version 1.0
 * @date 2021/9/16 10:48 上午
 * @desc
 */
@Slf4j
@Controller
@RequestMapping("/common")
public class CommonController {
    /**
     * 通用下载请求
     *
     * @param fileName 文件名称
     * @param delete   是否删除
    */
    @GetMapping("/download")
    @PreAuthorize("hasAuthority('sys:file:download') and (!#delete or hasAuthority('sys:file:delete'))")
    public void fileDownload(String fileName, Integer type, Boolean delete, HttpServletResponse response) {
        try {
            if (!FileUtils.checkAllowDownload(fileName)) {
                throw new BusinessException(ResponseCode.CODE_SYS_ERROR, "文件名称非法，不允许下载");
            }
            String realFileName = System.currentTimeMillis() + fileName.substring(fileName.indexOf("_") + 1);
            String filePath = getFilePath(fileName, type);

            response.setContentType(MediaType.APPLICATION_OCTET_STREAM_VALUE);
            FileUtils.setAttachmentResponseHeader(response, realFileName);
            FileUtils.writeBytes(filePath, response.getOutputStream());
            if (Boolean.TRUE.equals(delete)) {
                FileUtils.deleteFile(filePath);
            }
        } catch (BusinessException e) {
            throw e;
        } catch (Exception e) {
            log.error("下载文件失败", e);
            throw new BusinessException(ResponseCode.CODE_SYS_ERROR, "下载文件失败");
        }
    }

    private String getFilePath(String fileName, Integer type) {
        String filePath;
        if (type == DownloadType.LOCAL) {//本地资源下载
            filePath = SysConfig.getLocalStore() + "/" + fileName;
        } else {
            filePath = SysConfig.getDownloadPath() + "/" + fileName;
        }
        return filePath;
    }

    /**
     * 通用上传请求
     */
    @PostMapping("/upload")
    @ResponseBody
    @PreAuthorize("hasAuthority('sys:file:upload')")
    public JsonResponse uploadFile(MultipartFile file) {
        try {
            // 上传文件路径
            String filePath = SysConfig.getUploadPath();
            // 上传并返回新文件名称
            String fileName = FileUploadUtils.upload(filePath, file);

            return JsonResponse.success(fileName);
        } catch (Exception e) {
            return JsonResponse.error(ResponseCode.CODE_SYS_ERROR, e.getMessage());
        }
    }



    /**
     * 本地资源通用下载
    */
    @GetMapping("/download/resource2")
    @PreAuthorize("hasAuthority('sys:file:download')")
    public void resourceDownloadTest(String resource, HttpServletResponse response) {
        try {
            if (!FileUtils.checkAllowDownload(resource)) {
                throw new BusinessException(ResponseCode.CODE_SYS_ERROR, "文件名称非法，不允许下载");
            }
            // 本地资源路径
            String localPath = SysConfig.getDownloadPath() + "/" + resource;
            // 下载名称
            String downloadName = StringUtils.substringAfterLast(localPath, "/");
            response.setContentType(MediaType.APPLICATION_OCTET_STREAM_VALUE);
            FileUtils.setAttachmentResponseHeader(response, downloadName);
            FileUtils.writeBytes(localPath, response.getOutputStream());
        } catch (BusinessException e) {
            throw e;
        } catch (Exception e) {
            log.error("下载文件失败", e);
            throw new BusinessException(ResponseCode.CODE_SYS_ERROR, "下载文件失败");
        }
    }
}